refactor(gliner-guard): switch to bfloat16, cleanup formatting
- float16 → bfloat16 for better numerical stability on modern GPUs
- Code formatting cleanup (trailing spaces, blank lines)

+++ litserve-baseline/main.py
@@ -2,12 +2,14 @@ import litserve as ls
 from gliner2 import GLiNER2
 from pydantic import BaseModel
 import logging
-import os 
+import os
 from dotenv import load_dotenv
+
 load_dotenv()
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 import torch
+
 PII_LABELS = ["person", "address", "email", "phone"]
 SAFETY_LABELS = ["safe", "unsafe"]
 
@@ -19,7 +21,7 @@ class GuardRequest(BaseModel):
 class GLiNERGuardAPI(ls.LitAPI):
     def setup(self, device):
         self.model = GLiNER2.from_pretrained("hivetrace/gliner-guard-uniencoder")
-        self.model.to(device).to(torch.float16)
+        self.model.to(device).to(torch.bfloat16)
         self.schema = (
             self.model.create_schema()
             .entities(entity_types=PII_LABELS, threshold=0.4)
@@ -39,7 +41,6 @@ class GLiNERGuardAPI(ls.LitAPI):
             texts=batch,
             schemas=self.schema,
             batch_size=len(batch),
-            
         )
         return results
 
@@ -52,11 +53,11 @@ class GLiNERGuardAPI(ls.LitAPI):
 
 if __name__ == "__main__":
     api = GLiNERGuardAPI(max_batch_size=64, batch_timeout=0.05)
-    server = ls.LitServer(api, 
-                          accelerator="auto", 
-                          timeout=30, 
-                          workers_per_device=4,
-                          fast_queue=True,
-                          
-                          )
+    server = ls.LitServer(
+        api,
+        accelerator="auto",
+        timeout=30,
+        workers_per_device=4,
+        fast_queue=True,
+    )
     server.run(port=8000, generate_client_file=False)
